warning

+++ Classes/BITSessionStateData.m
@@ -43,7 +43,7 @@
   [super encodeWithCoder:coder];
   [coder encodeObject:self.envelopeTypeName forKey:@"envelopeTypeName"];
   [coder encodeObject:self.dataTypeName forKey:@"dataTypeName"];
-  [coder encodeInt:self.state forKey:@"self.state"];
+  [coder encodeInt:(int)self.state forKey:@"self.state"];
 }
 
 @end
